# GeoTIFF grid reading: perf improvements (fixes #2785)